+++ emperor/support_files/emperor/js/emperor.js
@@ -44,6 +44,18 @@ var g_visiblePoints = 0;
 var g_sphereScaler = 1.0;
 var g_keyBuilt = false;
 var g_useDiscreteColors = true;
+var g_screenshotBind;
+
+// valid ascii codes for filename
+var g_validAsciiCodes = new Array();
+// adding: .-_
+g_validAsciiCodes = g_validAsciiCodes.concat([45,46,95]);
+// adding: 0->9
+g_validAsciiCodes = g_validAsciiCodes.concat([48,49,50,51,52,53,54,55,56,57]);
+// adding: A->Z
+g_validAsciiCodes = g_validAsciiCodes.concat([65,66,67,68,68,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90]);
+// adding: a->z
+g_validAsciiCodes = g_validAsciiCodes.concat([97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118,119,120,121,122]);
 
 // taken from the qiime/colors.py module; a total of 29 colors
 k_QIIME_COLORS = [
@@ -71,7 +83,6 @@ k_QIIME_COLORS = [
 "0xA54700", // orange3
 "0x808000", // green3
 "0x008080"] // teal3
-
 
 // Taken from http://stackoverflow.com/questions/18082/validate-numbers-in-javascript-isnumeric
 function isNumeric(n) {
@@ -395,7 +406,7 @@ function colorByMenuChanged() {
 	colors = getColorList(vals);
 	
 	// build the colorby table in HTML
-	var lines = "<table>";
+	var lines = "<table id='colorbylist_table'>";
 	for(var i in vals){
 		// each field is identified by the value it has in the deduplicated
 		// list of values and by the number of the column in the mapping file
@@ -1595,20 +1606,77 @@ function drawEdges(){
 	}
 }
 
-function saveSVG(){
-	open("data:image/svg+xml," + encodeURIComponent(document.getElementById('main_plot').innerHTML));
-}
+/*Save the current view to SVG
+  This will take the current webGL renderer, convert it to SVG and then generate 
+  a file to download. Additionally it will create the labels if this option is selected.
+*/
+function saveSVG(button){
+    // add a name subfix for the filenames
+    if ((g_segments<=8 && g_visiblePoints>=10000) || (g_segments>8 && g_visiblePoints>=5000)) {
+        var res = confirm("The number of segments (" + g_segments + ") combined with the number " +
+            "of samples could take a long time and in some computers the browser will crash. " +
+            "If this happens we suggest to lower the number of segments or use the png " +
+            "implementation. Do you want to continue?");
+        if (res==false) return;
+    }
+    $('body').css('cursor','progress');
+    
+    var width = document.getElementById('pcoaPlotWrapper').offsetWidth;
+    var height = document.getElementById('pcoaPlotWrapper').offsetHeight;
+    
+    var color = $("#rendererbackgroundcolor").spectrum("get").toHexString(true);
+    var rendererBackgroundColor = new THREE.Color();
+    rendererBackgroundColor.setHex(color.replace('#','0x'));
 
-function SVGSaved(response){
-	var fileName = eval('('+response+')')
-	var dlwin = open('','Download SVG', 'width=400,height=50')
-
-	dlwin.document.open();
-	dlwin.document.write('<HTML><HEAD><meta name="content-disposition" content="inline; filename=\''+fileName+'\'">');
-	dlwin.document.write('</HEAD><BODY>');
-	dlwin.document.write('<a href=\''+fileName+'\'>'+fileName+'</a>')
-	dlwin.document.write('</BODY></HTML>');
-	dlwin.document.close();
+	var svgRenderer = new THREE.SVGRenderer({ antialias: true, preserveDrawingBuffer: true }); 
+	// this is the proper way to set the color of the background but it doesn't work   
+    svgRenderer.setClearColor(rendererBackgroundColor, 1);
+    svgRenderer.setSize(width, height);
+    svgRenderer.render(g_mainScene, g_sceneCamera);
+    svgRenderer.sortObjects = true;
+        
+    // converting svgRenderer to string: http://stackoverflow.com/questions/17398134/three-svgrenderer-save-text-of-image
+    var XMLS = new XMLSerializer(); 
+    var svgfile = XMLS.serializeToString(svgRenderer.domElement);
+    
+    // hacking the color to the svg
+    var index = svgfile.indexOf('viewBox="')+9;
+    var viewBox = svgfile.substring(index, svgfile.indexOf('"',index))
+    viewBox = viewBox.split(" ");
+    var background = '<rect id="background" height="' + viewBox[3] + '" width="' + viewBox[2] + '" y="' + 
+        viewBox[1] + '" x="' + viewBox[0] + '" stroke-width="0" stroke="#000000" fill="' + color + '"/>'
+    index = svgfile.indexOf('>',index)+1;
+    svgfile = svgfile.substr(0, index) + background + svgfile.substr(index);
+    
+    // adding xmlns header to open in the browser 
+    svgfile = svgfile.replace('viewBox=', 'xmlns="http://www.w3.org/2000/svg" viewBox=')
+    saveAs(new Blob([svgfile], {type: "text/plain;charset=utf-8"}), 
+         $('#saveas_name').val() + ".svg");
+    
+    if ($('#saveas_legends').is(':checked')) {
+        var labels_text = '', pos_y = 1, increment = 40, max_len = 0, font_size = 12;
+        $('#colorbylist_table tr div').each(function() {
+            if ($(this).attr('name').length > max_len) max_len = $(this).attr('name').length 
+            
+            // adding rectangle
+            labels_text += '<rect height="27" width="27" y="' + pos_y + 
+                '" x="1" stroke-width="1" ' + 'stroke="#FFFFFF" fill="' + 
+                $("#" + $(this).attr('id')).spectrum("get").toHexString(true) + '"/>';
+            // adding text
+            labels_text += '<text xml:space="preserve" y="' + (pos_y+20) + '" x="35" ' + 
+                'font-family="Monospace" font-size="' + font_size + '" stroke-width="0" ' +
+                'stroke="#000000" fill="#000000">' + $(this).attr('name') + '</text>';
+            pos_y += increment;
+        });
+        labels_text = '<svg width="' + ((font_size*max_len) + 10) + '" height="' + 
+            (pos_y-10) + '" xmlns="http://www.w3.org/2000/svg"><g>' + labels_text + 
+            '</g></svg>';
+        
+        saveAs(new Blob([labels_text], {type: "text/plain;charset=utf-8"}), 
+            $('#saveas_name').val() + "_labels.svg");
+    }
+    
+    $('body').css('cursor','default');
 }
 
 /*Utility function to draw two-vertices lines at a time
@@ -1776,6 +1844,9 @@ function togglePlots() {
 
 		// make all tabs usable
 		$("#menutabs").tabs({disabled: []});
+		
+		// adding ctrl-p
+		g_screenshotBind = THREEx.Screenshot.bindKey(g_mainRenderer, {charCode: 16});
 	}
 	// changes for parallel plots
 	else{
@@ -1791,7 +1862,12 @@ function togglePlots() {
 
 		// make the visibility, scaling, labels and axes tabs un-usable
 		// they have no contextualized meaning in when lookin at parallel plots
-		$("#menutabs").tabs({disabled: [2,3,4,5]});
+		// 0 = Key, 1 = Colors, 2 = Visibility, 3 = Scaling, 4 = Labels, 5 = Axes, 6 = View, 7 = Options
+		$("#menutabs").tabs({disabled: [2,3,4,5,7]});
+		
+		// removing the ctrl-p 
+        g_screenshotBind.unbind();
+		
 		colorByMenuChanged();
 	}
 }
@@ -1827,6 +1903,7 @@ function setParallelPlots() {
 $(document).ready(function() {
 	setJqueryUi()
 	
+	
 	// Detecting that webgl is activated
 	if ( ! Detector.webgl ) Detector.addGetWebGLMessage();
 
@@ -1843,6 +1920,13 @@ $(document).ready(function() {
 		g_sceneCamera.aspect = winAspect;
 		g_sceneCamera.updateProjectionMatrix();
 	});
+	
+	// Validating the string for the saveas filename = taken from http://stackoverflow.com/questions/6741175/trim-input-field-value-to-only-alphanumeric-characters-separate-spaces-with-wi
+    $('#saveas_name').keypress(function(event) {
+        var code = (event.keyCode ? event.keyCode : event.which);
+        if (g_validAsciiCodes.indexOf(code)==-1)
+            event.preventDefault();
+    });
 
 	init();
 	animate();
@@ -1935,10 +2019,13 @@ $(document).ready(function() {
 		rendererBackgroundColor.setHex("0x000000");
 
 		// renderer, the default background color is black
-		g_mainRenderer = new THREE.WebGLRenderer({ antialias: true });
+		g_mainRenderer = new THREE.WebGLRenderer({ antialias: true, preserveDrawingBuffer: true });
+        
+        // adding 'ctrl+p' to print screenshot
+        g_screenshotBind = THREEx.Screenshot.bindKey(g_mainRenderer, {charCode: 16});
+        
 		g_mainRenderer.setClearColor(rendererBackgroundColor, 1);
 		g_mainRenderer.setSize( document.getElementById('pcoaPlotWrapper').offsetWidth, document.getElementById('pcoaPlotWrapper').offsetHeight );
-		// g_mainRenderer.setSize( document.getElementById('vizualizations').offsetWidth , document.getElementById('vizualizations').offsetHeight );
 		g_mainRenderer.sortObjects = true;
 		main_plot.append(g_mainRenderer.domElement);