+++ ChangeLog.md
@@ -14,7 +14,10 @@ Emperor 0.9.1 (changes since Emperor 0.9.0 go here)
 * The "Colors" tab now has a selector, which allows to use the arrows to move between categories.
 * Default coloring scheme is discrete.
 * Add color pickers for the axes and axes labels.
+* To take a screenshot (PNG) of your current visualization you can press `ctrl+p`.
+* Export to SVG your visualization.
 * Emperor now relies on QIIME 1.7.0.
+* Added option `--number_of_segments` to control the quality of all spheres
 * Add color pickers for connecting bars in coordinate comparison plots.
 * Add option to select a master set of coordinates when making a comparison plot.